0.0.1 Prototype working

+++ InteractiveTransition/EPPZRandom.m
@@ -17,7 +17,7 @@
 
 float randomFloat()
 {
-    int size = NSIntegerMax;
+    int size = (int)NSIntegerMax;
     return (float)(arc4random() % size) / (float)size;
 }
